Generously generate QualifiedId for all package/r permutations.
My understanding of whats going on is it is possible for R files
in two different modules to generate the same id for different resoureces.

Butterknife used to use a Map<Integer,Id> where Id is a class that
contains knowledge of how to generate code. Thus, when ids are
duplicated only one would win the map. The loser would have it's
generated code import the wrong id.

Now we qualify the id with the package name of the id element. However,
due to the structure of the code-gen, we parse R files without knowing
exactly what packages reference which generated ids.

But we do know all the packages that reference something in the R file.
So we can defensively generated id's for ALL permutations. This sorta
stinks from a performance stand-point, but is the only quick way I can
figure out to use the QualifiedId, and gaurentee we generate all the
qualified ids we will need.

 Essentially, for every package
that has a butterknife annotation, we parse all R files in that
package with all package combinatios. This creates many
unneeded qualified ids, but it gaurentees that everything will be
present.